requirements.txt
flask

hikari
hikari[speedups]
hikari-crescent

python-dotenv

uvloop
orjson